init_backtrace()
